LPS-88715 Extract testGetAdaptiveMedias to a separate test class
